# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "提供用户积分管理、积分配置和积分日志查询等功能",
        "title": "积分管理系统API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/adminGrantPoints": {
            "post": {
                "description": "管理员给用户增加或减少积分\n获取当前登录用户的信息和积分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "用户管理",
                    "用户管理"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "description": "积分操作信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/configurePoints": {
            "post": {
                "description": "为文件设置积分值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分配置"
                ],
                "summary": "配置文件积分",
                "parameters": [
                    {
                        "description": "积分配置信息",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PointConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/downloadFile": {
            "post": {
                "description": "扣除用户积分并返回文件下载地址",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件下载"
                ],
                "summary": "下载文件",
                "parameters": [
                    {
                        "description": "文件信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getFileInfo": {
            "get": {
                "description": "获取文件名称和所需积分",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件下载"
                ],
                "summary": "获取文件信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件URL",
                        "name": "fileUrl",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getPointsList": {
            "get": {
                "description": "获取所有文件的积分配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分配置"
                ],
                "summary": "获取积分配置列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getPointsLog": {
            "get": {
                "description": "获取用户积分变更历史记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "积分日志"
                ],
                "summary": "获取积分变更日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名（可选）",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "返回记录数量限制（默认50）",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getUserInfo": {
            "get": {
                "description": "管理员给用户增加或减少积分\n获取当前登录用户的信息和积分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "tags": [
                    "用户管理",
                    "用户管理"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "description": "积分操作信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getUserPoints": {
            "get": {
                "description": "获取指定用户的积分信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户积分"
                ],
                "summary": "获取用户积分",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/getUsersList": {
            "get": {
                "description": "获取所有用户及其积分信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.PointConfig": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "积分描述",
                    "type": "string"
                },
                "fileUrl": {
                    "description": "文件路径",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "points": {
                    "description": "积分值",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}